Reduce memory footprint of OjbectId
By storing the 20 bytes of the SHA1 hash in an ObjectId
as an int and two longs, instead of a byte[20], the memory
footprint of a single ObjectId is reduced from ~56 bytes
to ~32 bytes.

+++ src/api/src/main/java/org/locationtech/geogig/model/RevObjects.java
@@ -51,9 +51,9 @@ public class RevObjects {
         Preconditions.checkArgument(numBytes > 0 && numBytes <= ObjectId.NUM_BYTES);
 
         StringBuilder sb = target == null ? new StringBuilder(2 * numBytes) : target;
-        byte b;
+        int b;
         for (int i = 0; i < numBytes; i++) {
-            b = (byte) id.byteN(i);
+            b = id.byteN(i);
             sb.append(HEX_DIGITS[(b >> 4) & 0xf]).append(HEX_DIGITS[b & 0xf]);
         }
         return sb;
